% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{checkDates}
\alias{checkDates}
\title{Checks the dates of a date_char column in a data.table}
\usage{
checkDates(dt)
}
\arguments{
\item{dt}{the data.table}
}
\description{
\code{checkDates} takes a data.table and splits the `date_char` column on either `/` or `-` (tries each).
}
\details{
Expects `date_char` to have the format x-x-x or x/x/x

Puts the results into 3 new columns () and returns the dt.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
