% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{fixAmbiguousDates}
\alias{fixAmbiguousDates}
\title{Fixes ambiguous dates}
\usage{
fixAmbiguousDates(dt)
}
\arguments{
\item{dt}{the data.table}
}
\description{
\code{fixAmbiguousDates} takes a data.table, fixes dates in dateFormat column and returns the dt.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
