% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{fixCircuitLabels_rf_24}
\alias{fixCircuitLabels_rf_24}
\title{Fixes circuit labels in rf_24}
\usage{
fixCircuitLabels_rf_24(dt)
}
\arguments{
\item{dt}{wide-form (original) datat table to fix & return}
}
\description{
\code{fixCircuitLabels_rf_24} is a stub that could be used to fix circuit labels.
    rf_24 has an additional circuit in some files but value is always NA so we ignore them (no fix).
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
