% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{fixCircuitLabels_rf_46}
\alias{fixCircuitLabels_rf_46}
\title{Fixes circuit labels in rf_46}
\usage{
fixCircuitLabels_rf_46(dt)
}
\arguments{
\item{dt}{wide-form (original) datat table to fix & return}
}
\description{
\code{fixCircuitLabels_rf_46} is a stub that could be used to fix circuit labels.
    rf_46 has 3 different versions of the circuit labels:
    1: Heat Pumps (2x) & Power$4232, Heat Pumps (2x) & Power$4399, Hot Water - Controlled$4231, Hot Water - Controlled$4400, Incomer - Uncontrolled$4230, Incomer - Uncontrolled$4401, Incomer Voltage$4405, Kitchen & Bedrooms$4229, Kitchen & Bedrooms$4402, Laundry & Bedrooms$4228, Laundry & Bedrooms$4403, Lighting$4233, Lighting$4404
    2: Heat Pumps (2x) & Power1$4232, Heat Pumps (2x) & Power2$4399, Hot Water - Controlled1$4231, Hot Water - Controlled2$4400, Incomer - Uncontrolled1$4230, Incomer - Uncontrolled2$4401, Incomer Voltage$4405, Kitchen & Bedrooms1$4229, Kitchen & Bedrooms2$4402, Laundry & Bedrooms1$4228, Laundry & Bedrooms2$4403, Lighting1$4233, Lighting2$4404
    3: Heat Pumps (2x) & Power_Imag$4399, Heat Pumps (2x) & Power$4232, Hot Water - Controlled_Imag$4400, Hot Water - Controlled$4231, Incomer - Uncontrolled_Imag$4401, Incomer - Uncontrolled$4230, Incomer Voltage$4405, Kitchen & Bedrooms_Imag$4402, Kitchen & Bedrooms$4229, Laundry & Bedrooms_Imag$4403, Laundry & Bedrooms$4228, Lighting_Imag$4404, Lighting$4233
    
    Was it re-used 3 times? Probably not as the labels are almost identical.
    
    This function fixes the labels to just the first (might also fix duplication of observations)
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
