% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{getCleanGridSpyFile}
\alias{getCleanGridSpyFile}
\title{Loads cleaned grid spy power data into a data.table}
\usage{
getCleanGridSpyFile(f)
}
\arguments{
\item{f}{file to load}
}
\description{
\code{getCleanGsFile}
}
\details{
Loads a clean Grid Spy data file using \code{readr::read_csv}. Ths enables .gz files to be autoloaded.

We STRONLY advise that you use this function to load the cleaned Grid Spy files as it:
 
 - forces dateTime_orig to be col_character to prevent readr attempting to parse it (incorrectly).
 - forces powerW be col_double to stop readr assuming an integer from parsing the first few lines
  
 We allow readr to auto-parse the r_dateTime column but note that this will set the timezone etc to the
 timezone correct for the current location. This may not be what you intend.
 
 Note that you can also use this function to load extracts from the Grid Spy files. It will just
 complain (throw a warning) if any of the columns it expects are missing.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
