% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{getGridSpyFileList}
\alias{getGridSpyFileList}
\title{Gets a list of gridSpy data files matching a pattern}
\usage{
getGridSpyFileList(fpath, pattern, mf)
}
\arguments{
\item{fpath}{where to look (recursively)}

\item{pattern}{the pattern to match}

\item{mf}{name of file to save interim file list and meta data to}
}
\description{
\code{getGridSpyFileList} takes a file path, searches for files matching pattern and returns the filelist as a data.table with metadata.
}
\details{
Also saves out the metadata as an interim file
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
