% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdData.R
\name{getHouseholdData}
\alias{getHouseholdData}
\title{Loads gridSpy household data from .xlsx}
\usage{
getHouseholdData(f)
}
\arguments{
\item{f}{the .xlsx file}
}
\description{
\code{getHouseholdData} does what it says. Expects .xlsx with two sheets, one for each sample.
}
\details{
Combines the results into one table. Could be extended to include other survey data when available.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
