% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getRunDateTime}
\alias{getRunDateTime}
\title{Get dateTime}
\usage{
getRunDateTime()
}
\description{
\code{getRunDateTime} returns a formated run dateTime.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
