% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tud.R
\name{getTUD}
\alias{getTUD}
\title{Loads time use data}
\usage{
getTUD(f)
}
\arguments{
\item{f}{the file}
}
\description{
\code{getTUD} does what it says - loads and cleans time-use diary data.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
