% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gzipIt}
\alias{gzipIt}
\title{Gzip a file}
\usage{
gzipIt(file)
}
\arguments{
\item{file}{file to gzip}
}
\description{
\code{gzipIt} gzips a file, over-writing automatically.
}
\author{
Michael Harper

Ben Anderson, \email{banderson@soton.ac.uk}
}
\concept{Utilities}
