% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{loadCleanGridSpyData}
\alias{loadCleanGridSpyData}
\title{Loads cleaned grid spy power data for all households between two dates into a single data.table}
\usage{
loadCleanGridSpyData(fPath, dateFrom, dateTo)
}
\arguments{
\item{fPath}{location of files to load}

\item{dateFrom}{date to start extract (inclusive)}

\item{dateTo}{date to end extract (inclusive)}
}
\description{
\code{loadCleanGridSpyData} checks to see if the extract file already exists. If not it loops over a file list and loads each in turn using \code{readr::read_csv}. It filters each file to
extract data between two dates and creates some derived time/date variables before
using \code{rbind} to create a single data.table which is saved and returned.
}
\details{
Function selects observations between
 \code{dateFrom} and \code{dateTo}.

Use of \code{readr::read_csv} enables .gz files to be autoloaded and proper parsing of dateTimes.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
