% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{processGridSpyDataFiles}
\alias{processGridSpyDataFiles}
\title{Loads and processes a list of gridSpy data files}
\usage{
processGridSpyDataFiles(dt, mf)
}
\arguments{
\item{dt}{list of files to load}

\item{mf}{name of file to save interim file list and meta data to}
}
\description{
\code{processGridSpyDataFiles} takes a file list (as a dt with attributes) and loads & processes data before saving out 1 file per household.
}
\details{
Updates the global file list and saves out the updated metadata
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
