% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Sets a range of parameters re-used throughout the repo by sourcing the <repo>/ggParams.R file}
\usage{
setup()
}
\description{
\code{setup} sources the <repo>/ggParams.R file which sets a range of parameters as elements of the `ggParams` list. These include data sources and .Rmd include locations.
}
\details{
The parameters can be over-written in scipts if needed but do so wih care!
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
