% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{fixCircuitLabels_rf_46}
\alias{fixCircuitLabels_rf_46}
\title{Fixes circuit labels in rf_46}
\usage{
fixCircuitLabels_rf_46(dt)
}
\arguments{
\item{dt}{wide-form (original) datat table to fix & return}
}
\description{
\code{fixCircuitLabels_rf_46} is used to fix circuit labels.
}
\details{
Inspection of the raw data shows that rf_46 has several different label sets.
    
    rf_46 was not re-used and the similarity of the label-sets seems to indicate typos.

It appears that power, imaginary power and voltage were all meansured in this household. But the labels are
    confused. See https://github.com/CfSOtago/GREENGridData/issues/1#issuecomment-520690467

This function does nothing (yet). For now it just passes the circuit labels back into the 'clean' data.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
