% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdData.R
\name{getHouseholdData}
\alias{getHouseholdData}
\title{Loads gridSpy household data from .xlsx}
\usage{
getHouseholdData(f)
}
\arguments{
\item{f}{the .xlsx file}
}
\description{
\code{getHouseholdData} does what it says. Expects an .xlsx with two sheets, one for each sample.
}
\details{
Combines & returns the results as a data.table which records:

- power company
   - location
   - n adults
   - n children aged 0-12
   - n chidlren aged 13-18
   - any recruitment or sampling notes
   - date gridSpy monitor disconnected and (potentially) re-used
   
   Will be extended to include other survey data when available.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
