% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSpyData.R
\name{processHhGridSpyData}
\alias{processHhGridSpyData}
\title{Loads and processes a list of gridSpy data files for a given household}
\usage{
processHhGridSpyData(hh, fileList)
}
\arguments{
\item{hh}{the household id}

\item{fileList}{list of files to load}
}
\description{
\code{getHhGridSpyData} takes a file list and loads & processes data before saving out 1 file. 
   It assumes all data is for one household (can only be detected from input file names)
}
\details{
Updates the per input file metadata and the per household per day summary stats file
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
